chore: Refactor menu folder functions into separate files
- Fixes #4
